Fix cargo clippy issues (#196)
* Fix warnings

* Fix more warnings

+++ src/api/method/get_multiple_new_address_proofs.rs
@@ -88,7 +88,7 @@ pub async fn get_multiple_new_address_proofs_helper(
             proof: proof.proof,
             lowElementLeafIndex: model.leaf_index as u32,
             merkleTree: tree,
-            rootSeq: proof.rootSeq as u64,
+            rootSeq: proof.rootSeq,
         };
         new_address_proofs.push(new_address_proof);
     }
